Leave callbacks alone
